D3D11-only experimental way to cull out-of-bounds triangles. Helps #5001, at least for Toca.
Really not sure how big the guard band should be, and the right way to
do this would be to use a geometry shader instead of killing triangles
by setting w = NaN, which I'm not sure will do sane things on all
hardware. Unfortunately geometry shaders are not available everywhere.